Extract mocks to separate files
